+++ CHANGELOG.md
@@ -14,10 +14,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - changed travis to use pylint on python 3.7 as well
+- changed the password in all valid functions to be optional
+- changed the examples to use random.SystemRandom instead of secrets module
+- added support for python 2.7
 
 ### Fixed
 
-- fixed a link in the README pointing to the old repo
+- fixed a link in the README pointing to the old repository
 
 ## [0.2.0] - 2018-08-07